Manjaro's Archived Forums certificate has expired
Right on the heels of rushing that Asahi kernel out the door, the skids have done it again. Like the changing of the seasons, you can almost count on Manjaro to screw something up as they do it so predictably & with such regularity! It's the one thing they manage to do consistently. They do not fuck up fucking up.

https://crt.sh/?q=66%3A16%3AD9%3A94%3AF4%3A7E%3A0F%3A87%3A1B%3A95%3A83%3A7D%3A7A%3ADE%3AB5%3A41%3A05%3A84%3A95%3A56